agent/eclipse_zero.py
#!/usr/bin/env python3
"""
Eclipse Zero - Unified AI Agent
Merges Eclipse tool-based execution with Agent Zero autonomous patterns
"""

import asyncio
import json
import os
import sys
from datetime import datetime
from typing import Any, Dict, List, Optional
from dataclasses import dataclass, field
from enum import Enum
from pathlib import Path

import uvicorn
import yaml
from fastapi import FastAPI, WebSocket, WebSocketDisconnect
from fastapi.middleware.cors import CORSMiddleware

# Load configuration
CONFIG_PATH = Path(os.getenv("CONFIG_PATH", "config"))
MEMORY_PATH = Path(os.getenv("MEMORY_PATH", "memory"))

class AgentState(Enum):
    IDLE = "idle"
    PLANNING = "planning"
    EXECUTING = "executing"
    DELEGATING = "delegating"
    COMPLETING = "completing"

@dataclass
class AgentContext:
    """Session context with memory"""
    session_id: str
    created_at: datetime = field(default_factory=datetime.now)
    messages: List[Dict] = field(default_factory=list)
    tools_used: List[str] = field(default_factory=list)
    sub_agents: List[str] = field(default_factory=list)
    memories_loaded: List[str] = field(default_factory=list)
    
    def add_message(self, role: str, content: str, metadata: Dict = None):
        self.messages.append({
            "role": role,
            "content": content,
            "timestamp": datetime.now().isoformat(),
            "metadata": metadata or {}
        })

class EclipseZero:
    """
    Unified agent combining Eclipse tools with Agent Zero patterns
    """
    
    def __init__(self):
        self.config = self._load_config()
        self.contexts: Dict[str, AgentContext] = {}
        self.state = AgentState.IDLE
        self.memory_engine = None  # Initialize in setup
        
    def _load_config(self) -> Dict:
        """Load agent configuration"""
        config_file = CONFIG_PATH / "agents.yaml"
        if config_file.exists():
            with open(config_file) as f:
                return yaml.safe_load(f)
        return {}
    
    async def process_task(self, session_id: str, task: str) -> Dict[str, Any]:
        """
        Process a task using the 4-step Agent Zero protocol:
        1. Outline Plan
        2. Check Resources (memories, skills)
        3. Break Down & Solve (tools/sub-agents)
        4. Complete & Verify
        """
        # Get or create context
        if session_id not in self.contexts:
            self.contexts[session_id] = AgentContext(session_id=session_id)
        
        ctx = self.contexts[session_id]
        ctx.add_message("user", task)
        
        # === Step 0: Outline Plan ===
        self.state = AgentState.PLANNING
        plan = await self._outline_plan(task)
        
        # === Step 1: Check Resources ===
        memories = await self._check_memories(task)
        skills = await self._check_skills(task)
        
        # === Step 2: Break Down & Solve ===
        self.state = AgentState.EXECUTING
        
        results = []
        if plan.get("requires_sub_agents", False):
            # Delegate to sub-agents in parallel
            self.state = AgentState.DELEGATING
            sub_tasks = plan.get("sub_tasks", [])
            sub_results = await self._delegate_sub_agents(session_id, sub_tasks)
            results.extend(sub_results)
        else:
            # Execute directly with tools
            tool_results = await self._execute_with_tools(session_id, plan)
            results.append(tool_results)
        
        # === Step 3: Complete & Verify ===
        self.state = AgentState.COMPLETING
        response = await self._synthesize_response(task, memories, results)
        
        # Save learnings
        await self._save_memory(task, response, ctx.tools_used)
        
        ctx.add_message("assistant", response["content"], response.get("metadata"))
        
        return {
            "session_id": session_id,
            "response": response["content"],
            "plan": plan,
            "memories_used": memories,
            "tools_used": ctx.tools_used,
            "state": self.state.value
        }
    
    async def _outline_plan(self, task: str) -> Dict:
        """Step 0: Understand what needs to be done"""
        # Simple planning logic - can be enhanced with LLM
        is_complex = len(task) > 100 or any(kw in task.lower() for kw in [
            "complex", "multiple", "analyze", "refactor", "large codebase"
        ])
        
        return {
            "task": task,
            "complexity": "complex" if is_complex else "simple",
            "requires_sub_agents": is_complex,
            "sub_tasks": self._break_into_subtasks(task) if is_complex else [],
            "estimated_steps": 3 if is_complex else 1
        }
    
    def _break_into_subtasks(self, task: str) -> List[str]:
        """Break complex task into subtasks"""
        # This would use LLM in production
        return [
            f"Analyze: {task[:50]}...",
            f"Execute: {task[:50]}...",
            f"Verify: {task[:50]}..."
        ]
    
    async def _check_memories(self, task: str) -> List[Dict]:
        """Step 1a: Check for relevant past work"""
        memories = []
        # Load from JSON memory store
        memory_dir = MEMORY_PATH / "json"
        if memory_dir.exists():
            for mem_file in sorted(memory_dir.glob("mem_*.json"), reverse=True)[:5]:
                try:
                    with open(mem_file) as f:
                        mem = json.load(f)
                        if any(kw in mem.get("content", "").lower() for kw in task.lower().split()[:3]):
                            memories.append(mem)
                except Exception:
                    continue
        return memories
    
    async def _check_skills(self, task: str) -> List[str]:
        """Step 1b: Check available skills"""
        skills = []
        skills_dir = Path("skills")
        if skills_dir.exists():
            for skill_file in skills_dir.rglob("SKILL.md"):
                skills.append(str(skill_file.parent.name))
        return skills
    
    async def _execute_with_tools(self, session_id: str, plan: Dict) -> Dict:
        """Execute task using available tools"""
        # Tool execution would be implemented here
        # For now, return placeholder
        return {
            "status": "success",
            "tools_used": ["Shell", "ReadFile"],
            "output": "Executed with tools"
        }
    
    async def _delegate_sub_agents(self, session_id: str, sub_tasks: List[str]) -> List[Dict]:
        """Delegate subtasks to parallel sub-agents"""
        # Parallel execution of sub-agents
        tasks = [self._run_sub_agent(session_id, i, subtask) for i, subtask in enumerate(sub_tasks)]
        results = await asyncio.gather(*tasks, return_exceptions=True)
        return [{"index": i, "result": r} for i, r in enumerate(results) if not isinstance(r, Exception)]
    
    async def _run_sub_agent(self, parent_session: str, index: int, subtask: str) -> Dict:
        """Run a single sub-agent task"""
        sub_session = f"{parent_session}_sub_{index}"
        # Sub-agent execution logic
        return {
            "sub_session": sub_session,
            "task": subtask,
            "status": "completed",
            "output": f"Completed: {subtask[:50]}..."
        }
    
    async def _synthesize_response(self, task: str, memories: List[Dict], results: List[Dict]) -> Dict:
        """Step 3: Combine results into final response"""
        # Response synthesis
        return {
            "content": f"Completed task: {task[:50]}...\n\nResults synthesized from {len(results)} execution steps.",
            "metadata": {
                "memories_referenced": len(memories),
                "execution_steps": len(results)
            }
        }
    
    async def _save_memory(self, task: str, response: Dict, tools_used: List[str]):
        """Save useful learnings to memory"""
        memory_dir = MEMORY_PATH / "json"
        memory_dir.mkdir(parents=True, exist_ok=True)
        
        memory_id = f"mem_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
        memory = {
            "id": memory_id,
            "content": f"Task: {task[:200]}...",
            "category": "execution",
            "tags": tools_used,
            "created_at": datetime.now().isoformat(),
            "access_count": 0
        }
        
        with open(memory_dir / f"{memory_id}.json", 'w') as f:
            json.dump(memory, f, indent=2)

# FastAPI Application
app = FastAPI(title="Eclipse Memory", version="1.0.0")
agent = EclipseZero()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/health")
async def health():
    return {"status": "healthy", "agent_state": agent.state.value}

@app.post("/chat")
async def chat(request: Dict):
    """Main chat endpoint"""
    session_id = request.get("session_id", f"session_{datetime.now().strftime('%Y%m%d_%H%M%S')}")
    task = request.get("message", "")
    
    result = await agent.process_task(session_id, task)
    return result

@app.get("/sessions/{session_id}")
async def get_session(session_id: str):
    """Get session context"""
    if session_id in agent.contexts:
        ctx = agent.contexts[session_id]
        return {
            "session_id": ctx.session_id,
            "created_at": ctx.created_at.isoformat(),
            "message_count": len(ctx.messages),
            "tools_used": ctx.tools_used
        }
    return {"error": "Session not found"}

@app.websocket("/ws/{session_id}")
async def websocket_endpoint(websocket: WebSocket, session_id: str):
    """WebSocket for real-time communication"""
    await websocket.accept()
    try:
        while True:
            data = await websocket.receive_json()
            result = await agent.process_task(session_id, data.get("message", ""))
            await websocket.send_json(result)
    except WebSocketDisconnect:
        print(f"Session {session_id} disconnected")

if __name__ == "__main__":
    port = int(os.getenv("PORT", 50001))
    uvicorn.run(app, host="0.0.0.0", port=port)
