requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# AI/LLM
openai>=1.6.0
anthropic>=0.8.0
litellm>=1.15.0
tiktoken>=0.5.2

# Memory
chromadb>=0.4.18
sqlite-vec>=0.0.1
redis>=5.0.1

# Tools
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
schedule>=1.2.0

# Development (optional)
pytest>=7.4.3
black>=23.11.0
mypy>=1.7.1
